+++ test/common/sharedkey-tests.js
@@ -48,7 +48,7 @@ describe('sharedkey-tests', function () {
     webResource.withHeader(HeaderConstants.MS_DATE, 'Fri, 23 Sep 2011 01:37:34 GMT');
 
     sharedkey.signRequest(webResource, function () {
-      assert.equal(webResource.headers[HeaderConstants.AUTHORIZATION], 'SharedKey devstoreaccount1:+BPqJb08vDu2qpM43bvpBbebyfVf5uZow+tOkqbbKWw=');
+      assert.equal(webResource.headers[HeaderConstants.AUTHORIZATION], 'SharedKey devstoreaccount1:phWLwBDQmLRuhOCKblR0+k6+6N/EpF236K5KID6VXx8=');
 
       done();
     });
@@ -80,7 +80,7 @@ describe('sharedkey-tests', function () {
     assert.equal(sasQueryString[QueryStringConstants.SIGNED_PROTOCOL], sharedAccessPolicy.AccessPolicy.Protocols);
     assert.equal(sasQueryString[QueryStringConstants.SIGNED_IP], sharedAccessPolicy.AccessPolicy.IPAddressOrRange);
     assert.equal(sasQueryString[QueryStringConstants.SIGNED_VERSION], HeaderConstants.TARGET_STORAGE_VERSION);
-    assert.equal(sasQueryString[QueryStringConstants.SIGNATURE], 'kb1m7JjHBcB2eXle9m/jzNXyLh3cCA2VvHWM+A9My8k=');
+    assert.equal(sasQueryString[QueryStringConstants.SIGNATURE], '4J76vq0uIO2nuRsgVPA2bJiKdKTmDfeghzumaLhOPas=');
 
     done();
   });