+++ test/services/blob/blobservice-tests.js
@@ -318,11 +318,11 @@ describe('BlobService', function () {
           assert.equal(uploadError, null);
           assert.ok(uploadResponse.isSuccessful);
 
-          blobService.getBlobToText(containerName, blobName, function (downloadErr, blobText, blob, downloadResponse) {
+          blobService.getBlobToText(containerName, blobName, function (downloadErr, content, blob, downloadResponse) {
             assert.equal(downloadErr, null);
             assert.ok(downloadResponse.isSuccessful);
             assert.ok(blob);
-            assert.equal(blobText, blobText);
+            assert.strictEqual(content, blobText);
 
             done();
           });
@@ -337,11 +337,11 @@ describe('BlobService', function () {
           assert.equal(uploadError, null);
           assert.ok(uploadResponse.isSuccessful);
 
-          blobService.getBlobToText(containerName, blobName, function (downloadErr, blobText, blob, downloadResponse) {
+          blobService.getBlobToText(containerName, blobName, function (downloadErr, content, blob, downloadResponse) {
             assert.equal(downloadErr, null);
             assert.ok(downloadResponse.isSuccessful);
             assert.ok(blob);
-            assert.equal(blobText, blobText);
+            assert.strictEqual(content, blobText);
 
             done();
           });
@@ -356,11 +356,11 @@ describe('BlobService', function () {
           assert.equal(uploadError, null);
           assert.ok(uploadResponse.isSuccessful);
 
-          blobService.getBlobToText(containerName, blobName, function (downloadErr, blobText, blob, downloadResponse) {
+          blobService.getBlobToText(containerName, blobName, function (downloadErr, content, blob, downloadResponse) {
             assert.equal(downloadErr, null);
             assert.ok(downloadResponse.isSuccessful);
             assert.ok(blob);
-            assert.equal(blobText, blobText);
+            assert.strictEqual(content, blobText);
 
             done();
           });
@@ -375,11 +375,11 @@ describe('BlobService', function () {
           assert.equal(uploadError, null);
           assert.ok(uploadResponse.isSuccessful);
 
-          blobService.getBlobToText(containerName, blobName, function (downloadErr,  blobText, blob, downloadResponse) {
+          blobService.getBlobToText(containerName, blobName, function (downloadErr, content, blob, downloadResponse) {
             assert.equal(downloadErr, null);
             assert.ok(downloadResponse.isSuccessful);
             assert.ok(blob);
-            assert.equal(blobText, blobText);
+            assert.equal(content, blobText);
 
             done();
           });
@@ -656,7 +656,7 @@ describe('BlobService', function () {
             assert.equal(getErr, null);
 
             assert.notEqual(blob, null);
-            assert.notEqual(blob.serverEncrypted, false);
+            assert.notEqual(blob.serverEncrypted, null); //Note the storage account for this test suite could have enabled or disabled SSE.
 
             if (blob) {
               assert.notEqual(blob.metadata, null);
@@ -1286,7 +1286,7 @@ describe('BlobService', function () {
         assert.strictEqual(parsedUrl.port, '80');
         assert.strictEqual(parsedUrl.hostname, 'host.com');
         assert.strictEqual(parsedUrl.pathname, '/' + containerName + '/' + blobName);
-        assert.strictEqual(parsedUrl.query, 'se=2011-10-12T11%3A53%3A40Z&spr=https&sv=2016-05-31&sr=b&sig=5ubgzWFxfpW857DpF5QVK9HNbewzuQHjvwB%2BlGEdubM%3D');
+        assert.strictEqual(parsedUrl.query, 'se=2011-10-12T11%3A53%3A40Z&spr=https&sv=2017-04-17&sr=b&sig=A7y9u890HRxG0XSEsRntrGJ7WIzwrXC5ttro2KFOySU%3D');
 
         blobUrl = blobServiceassert.getUrl(containerName, blobName, sasToken, false, '2016-10-11T11:03:40Z');
 
@@ -1295,7 +1295,7 @@ describe('BlobService', function () {
         assert.strictEqual(parsedUrl.port, '80');
         assert.strictEqual(parsedUrl.hostname, 'host-secondary.com');
         assert.strictEqual(parsedUrl.pathname, '/' + containerName + '/' + blobName);
-        assert.strictEqual(parsedUrl.query, 'se=2011-10-12T11%3A53%3A40Z&spr=https&sv=2016-05-31&sr=b&sig=5ubgzWFxfpW857DpF5QVK9HNbewzuQHjvwB%2BlGEdubM%3D&snapshot=2016-10-11T11%3A03%3A40Z');
+        assert.strictEqual(parsedUrl.query, 'se=2011-10-12T11%3A53%3A40Z&spr=https&sv=2017-04-17&sr=b&sig=A7y9u890HRxG0XSEsRntrGJ7WIzwrXC5ttro2KFOySU%3D&snapshot=2016-10-11T11%3A03%3A40Z');
 
         done();
       });
@@ -1376,7 +1376,7 @@ describe('BlobService', function () {
       assert.equal(sasQueryString[QueryStringConstants.SIGNED_PERMISSIONS], BlobUtilities.SharedAccessPermissions.READ);
       assert.equal(sasQueryString[QueryStringConstants.SIGNED_PROTOCOL], 'https');
       assert.equal(sasQueryString[QueryStringConstants.SIGNED_VERSION], HeaderConstants.TARGET_STORAGE_VERSION);
-      assert.equal(sasQueryString[QueryStringConstants.SIGNATURE], 'JM+OTBtD7HFVD3A/I5r/iE0HlW8yLcv7DMQoqDT/rtw=');
+      assert.equal(sasQueryString[QueryStringConstants.SIGNATURE], 'Qa15RHcYQIkyLGqQgvf5qbHesBnK5Uzjgzzpo92vCXc=');
 
       done();
     });