+++ typings/azure-storage/azure-storage.d.ts
@@ -183,7 +183,31 @@ declare module azurestorage {
           */
           setServiceProperties(serviceProperties: common.models.ServicePropertiesResult.ServiceProperties, options: common.RequestOptions, callback: ErrorOrResponse): void;
           setServiceProperties(serviceProperties: common.models.ServicePropertiesResult.ServiceProperties, callback: ErrorOrResponse): void;
-          
+
+          /**
+          * Sets the tier of a pageblob under a premium storage account.
+          *
+          * @this {BlobService}
+          * @param {string}             container                                The container name.
+          * @param {string}             blob                                     The blob name.
+          * @param {string}             blobTier                                 Please see BlobUtilities.BlobTier.PremiumPageBlobTier for possible values.
+          * @param {LocationMode}       [options.locationMode]                   Specifies the location mode used to decide which location the request should be sent to. 
+          *                                                                      Please see StorageUtilities.LocationMode for the possible values.
+          * @param {int}                [options.timeoutIntervalInMs]            The server timeout interval, in milliseconds, to use for the request.
+          * @param {int}                [options.clientRequestTimeoutInMs]       The timeout of client requests, in milliseconds, to use for the request.
+          * @param {int}                [options.maximumExecutionTimeInMs]       The maximum execution time, in milliseconds, across all potential retries, to use when making this request.
+          *                                                                      The maximum execution time interval begins at the time that the client begins building the request. The maximum
+          *                                                                      execution time is checked intermittently while performing requests, and before executing retries.
+          * @param {string}             [options.clientRequestId]                A string that represents the client request ID with a 1KB character limit.
+          * @param {bool}               [options.useNagleAlgorithm]              Determines whether the Nagle algorithm is used; true to use the Nagle algorithm; otherwise, false.
+          *                                                                      The default value is false.
+          * @param {errorOrResponse}    callback                                 `error` will contain information
+          *                                                                      if an error occurs; otherwise, `response`
+          *                                                                      will contain information related to this operation.
+          */
+          setBlobTier(container: string, blob: string, blobTier: string, options: common.RequestOptions, callback: ErrorOrResponse): void;
+          setBlobTier(container: string, blob: string, blobTier: string, callback: ErrorOrResponse): void;
+
           /**
           * Lists a segment containing a collection of container items under the specified account.
           *
@@ -1387,6 +1411,7 @@ declare module azurestorage {
           * @param {string}             targetContainer                           The target container name.
           * @param {string}             targetBlob                                The target blob name.
           * @param {object}             [options]                                 The request options.
+          * @param {string}             [options.blobTier]                        For page blobs on premium accounts only. Set the tier of target blob. Refer to BlobUtilities.BlobTier.PremiumPageBlobTier.
           * @param {string}             [options.snapshotId]                      The source blob snapshot identifier.
           * @param {object}             [options.metadata]                        The target blob metadata key/value pairs.
           * @param {string}             [options.leaseId]                         The target blob lease identifier.
@@ -1514,6 +1539,7 @@ declare module azurestorage {
           * @param {bool}               [options.storeBlobContentMD5]                   Specifies whether the blob's ContentMD5 header should be set on uploads.
           *                                                                             The default value is false for page blobs.
           * @param {bool}               [options.useTransactionalMD5]                   Calculate and send/validate content MD5 for transactions.
+          * @param {string}             [options.blobTier]                              For page blobs on premium accounts only. Set the tier of the target blob. Refer to BlobUtilities.BlobTier.PremiumPageBlobTier.
           * @param {string}             [options.contentSettings.contentType]           The MIME content type of the blob. The default type is application/octet-stream.
           * @param {string}             [options.contentSettings.contentEncoding]       The content encodings that have been applied to the blob.
           * @param {string}             [options.contentSettings.contentLanguage]       The natural languages used by this resource.
@@ -1598,6 +1624,7 @@ declare module azurestorage {
           * @param {bool}               [options.storeBlobContentMD5]                   Specifies whether the blob's ContentMD5 header should be set on uploads.
           *                                                                             The default value is false for page blobs.
           * @param {bool}               [options.useTransactionalMD5]                   Calculate and send/validate content MD5 for transactions.
+          * @param {string}             [options.blobTier]                              For page blobs on premium accounts only. Set the tier of the target blob. Refer to BlobUtilities.BlobTier.PremiumPageBlobTier.
           * @param {string}             [options.contentSettings.contentType]           The MIME content type of the blob. The default type is application/octet-stream.
           * @param {string}             [options.contentSettings.contentEncoding]       The content encodings that have been applied to the blob.
           * @param {string}             [options.contentSettings.contentLanguage]       The natural languages used by this resource.
@@ -1623,6 +1650,7 @@ declare module azurestorage {
           /**
           * Provides a stream to write to a page blob. Assumes that the blob exists.
           * If it does not, please create the blob using createPageBlob before calling this method or use createWriteStreamNewPageBlob.
+          * Please note the `Stream` returned by this API should be used with piping.
           *
           * @this {BlobService}
           * @param {string}             container                                       The container name.
@@ -1669,6 +1697,7 @@ declare module azurestorage {
 
           /**
           * Provides a stream to write to a page blob. Creates the blob before writing data.
+          * Please note the `Stream` returned by this API should be used with piping.
           *
           * @this {BlobService}
           * @param {string}             container                                       The container name.
@@ -1682,6 +1711,7 @@ declare module azurestorage {
           * @param {bool}               [options.storeBlobContentMD5]                   Specifies whether the blob's ContentMD5 header should be set on uploads.
           *                                                                             The default value is false for page blobs and true for block blobs.
           * @param {bool}               [options.useTransactionalMD5]                   Calculate and send/validate content MD5 for transactions.
+          * @param {string}             [options.blobTier]                              For page blobs on premium accounts only. Set the tier of the target blob. Refer to BlobUtilities.BlobTier.PremiumPageBlobTier.
           * @param {string}             [options.contentSettings.contentType]           The MIME content type of the blob. The default type is application/octet-stream.
           * @param {string}             [options.contentSettings.contentEncoding]       The content encodings that have been applied to the blob.
           * @param {string}             [options.contentSettings.contentLanguage]       The natural languages used by this resource.
@@ -2039,6 +2069,7 @@ declare module azurestorage {
 
           /**
           * Provides a stream to write to a block blob.
+          * Please note the `Stream` returned by this API should be used with piping.
           *
           * @this {BlobService}
           * @param {string}             container                                       The container name.
@@ -2425,6 +2456,7 @@ declare module azurestorage {
           * To avoid overwriting and instead throw an error if the blob exists, please pass in an accessConditions parameter in the options object.
           * This API should be used strictly in a single writer scenario because the API internally uses the append-offset conditional header to avoid duplicate blocks.
           * If you are guaranteed to have a single writer scenario, please look at options.absorbConditionalErrorsOnRetry and see if setting this flag to true is acceptable for you.
+          * Please note the `Stream` returned by this API should be used with piping.
           *
           * @this {BlobService}
           * @param {string}             container                                     The container name.
@@ -2471,6 +2503,7 @@ declare module azurestorage {
           * If it does not, please create the blob using createAppendBlob before calling this method or use createWriteStreamToNewAppendBlob.
           * This API should be used strictly in a single writer scenario because the API internally uses the append-offset conditional header to avoid duplicate blocks.
           * If you are guaranteed to have a single writer scenario, please look at options.absorbConditionalErrorsOnRetry and see if setting this flag to true is acceptable for you.
+          * Please note the `Stream` returned by this API should be used with piping.
           *
           * @this {BlobService}
           * @param {string}             container                                     The container name.
@@ -2846,6 +2879,8 @@ declare module azurestorage {
             lastModified: string;
             contentLength: string;
             blobType: string;
+            accessTier?: string;
+            accessTierInferred?: boolean;
             isIncrementalCopy?: boolean;
             requestId: string;
             sequenceNumber?: string;
@@ -2884,6 +2919,7 @@ declare module azurestorage {
             metadata?: Object;
             leaseId?: string;
             transactionalContentMD5?: string;
+            blobTier?: string;
             contentSettings?: {
               contentType?: string;
               contentEncoding?: string;
@@ -3026,6 +3062,18 @@ declare module azurestorage {
             UPDATE: string;
             INCREMENT: string;
           };
+          BlobTier: {
+            PremiumPageBlobTier: {
+              P4: string;
+              P6: string;
+              P10: string;
+              P20: string;
+              P30: string;
+              P40: string;
+              P50: string;
+              P60: string;
+            };
+          }
         };
       }
     }
@@ -7126,6 +7174,7 @@ declare module azurestorage {
         /**
         * Provides a stream to write to a file. Assumes that the file exists.
         * If it does not, please create the file using createFile before calling this method or use createWriteStreamNewFile.
+        * Please note the `Stream` returned by this API should be used with piping.
         *
         * @this {FileService}
         * @param {string}             share                                         The share name.
@@ -7168,6 +7217,7 @@ declare module azurestorage {
 
         /**
         * Provides a stream to write to a file. Creates the file before writing data.
+        * Please note the `Stream` returned by this API should be used with piping.
         *
         * @this {FileService}
         * @param {string}             share                                         The share name.
@@ -7298,6 +7348,7 @@ declare module azurestorage {
           lastModified: string;
           requestId?: string;
           metadata?: { [key: string]: string; };
+          serverEncrypted?: string;
           exists?: boolean;
           created?: boolean;
         }
@@ -7313,6 +7364,7 @@ declare module azurestorage {
           contentRange: string;
           contentLength: string;
           metadata?: { [key: string]: string; };
+          serverEncrypted?: string;
           contentSettings?:{
             contentEncoding: string;
             contentLanguage: string;
@@ -7441,6 +7493,14 @@ declare module azurestorage {
       // ./services/file/fileutilities
       // ###########################
       export module FileUtilities {
+        var SharedAccessPermissions: {
+          READ: string;
+          CREATE: string;
+          WRITE: string;
+          DELETE: string;
+          LIST: string;
+        };
+
         var ListingDetails: {
           METADATA: string;
         };
@@ -9262,6 +9322,7 @@ declare module azurestorage {
     headers?: Map<string>;
     md5: string;
     error?: StorageError | Error;
+    requestServerEncrypted?: boolean;
   }
 
   interface ServiceResult {