+++ ChangeLog.md
@@ -4,7 +4,8 @@ be taken. This is a GA release and the changes described below indicate the chan
 2017.08 Version 2.3.0
 
 ALL
-* Updated storage service version to 2017-04-17. Fore more information, please see - https://docs.microsoft.com/en-us/rest/api/storageservices/versioning-for-the-azure-storage-services
+* Updated storage service version to 2017-04-17. For more information, please see - https://docs.microsoft.com/en-us/rest/api/storageservices/versioning-for-the-azure-storage-services
+* Updated the dependency of the 'request' module to avoid security vulnerability: (https://snyk.io/test/npm/azure-storage). 
 * Added `requestServerEncrypted` property to `ServiceResponse` which indicates if the contents of the request have been successfully encrypted.
 
 BLOB
@@ -12,6 +13,10 @@ BLOB
 
 FILE
 * Added `serverEncryped` property to `FileResult` and `DirectoryResult` which indicates if the file data and application metadata are completely encrypted using the specified algorithm on the server.
+* Fix a TypeScript issue that SharedAccessPermissions for FileUtilities is missing in type definition file.
+
+TABLE
+* Fix a typo in table query example for combineFilters function.
 
 2017.08 Version 2.2.2